refactor: KnowledgeSection GSAP ScrollTrigger 애니메이션 추가 (제목 페이드인, Lottie 튕김, 리스트 슬라이드인)
