+++ src/features/home/HeroSection.jsx
@@ -1,25 +1,143 @@
-import React, { Suspense, lazy } from 'react';
+import React, { useLayoutEffect, useRef } from 'react';
 import styled from '@emotion/styled';
-import Underlay from './Hero/Underlay';
-import LoadingSpinner from '@common/LoadingSpinner';
+import gsap from 'gsap';
+import { TeamHeroImages } from './utils/getTeamHeroImage';
+import bgImg from './assets/hero-bg.svg';
 
-const ThreeScene = lazy(() => import('./Hero/ThreeScene'));
-
-const HeroContainer = styled.div`
-  height: 100vh;
+const TeamHeroContainer = styled.div`
+  min-height: 100vh;
   position: relative;
-  background: linear-gradient(180deg, #e6eaf5 0%, #f6f6f6 100%);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  overflow: hidden;
 `;
 
-const HeroSection = () => {
-  return (
-    <HeroContainer data-testid="hero-section">
-      <Suspense fallback={<LoadingSpinner />}>
-        <ThreeScene />
-      </Suspense>
-      <Underlay />
-    </HeroContainer>
-  );
-};
+const BgImage = styled.img`
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  opacity: 0.5;
+`;
 
-export default HeroSection;
+const TextContainer = styled.div`
+  position: absolute;
+  text-align: center;
+  p {
+    font-family: 'IBM Plex Sans', sans-serif;
+    font-weight: 300;
+    font-size: 9rem;
+  }
+
+  span {
+    color: var(--primary);
+  }
+
+  @media (max-width: 768px) {
+    p {
+      font-size: 4rem;
+    }
+  }
+`;
+
+const ImageContainer = styled.div`
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+  place-items: center;
+  gap: 14rem;
+  width: 100%;
+
+  img {
+    width: 300px;
+    height: 300px;
+    object-fit: cover;
+  }
+
+  @media (max-width: 768px) {
+    grid-template-columns: repeat(2, 1fr);
+    gap: 14rem 4rem;
+
+    img {
+      width: 150px;
+      height: 150px;
+    }
+
+    img:nth-of-type(n + 5) {
+      display: none;
+    }
+  }
+`;
+
+export default function TeamHero() {
+  const sectionRef = useRef(null);
+  const textRef = useRef(null);
+
+  useLayoutEffect(() => {
+    const tl = gsap.timeline({
+      scrollTrigger: {
+        trigger: sectionRef.current,
+        start: 'top top',
+        end: 'bottom center',
+        pin: true,
+        scrub: 1,
+      },
+    });
+
+    gsap.utils.toArray('.img').forEach((img) => {
+      tl.to(
+        img,
+        {
+          x: random(-2000, 2000),
+          y: random(-1000, 1000),
+          rotate: random(-700, 700),
+          scale: 0,
+        },
+        0,
+      );
+    });
+    gsap.to(textRef.current, {
+      scale: 0,
+      scrollTrigger: {
+        start: () => tl.scrollTrigger.end,
+        end: '80%',
+        pin: textRef.current,
+        pinType: 'transform',
+        scrub: 1,
+      },
+    });
+    gsap.fromTo(
+      '.bg-image',
+      { opacity: 0.2 },
+      {
+        opacity: 0,
+        ease: 'none',
+        scrollTrigger: {
+          trigger: sectionRef.current,
+          start: 'top top',
+          end: 'bottom top',
+          scrub: 1,
+        },
+      },
+    );
+  }, []);
+
+  const random = (min, max) => min + (max - min) * Math.random();
+
+  return (
+    <TeamHeroContainer ref={sectionRef}>
+      <BgImage src={bgImg} alt="bg" className="bg-image" />
+      <TextContainer>
+        <p ref={textRef}>
+          'One a Day' <span>HARUHAN</span>
+        </p>
+      </TextContainer>
+      <ImageContainer>
+        {TeamHeroImages.map((src, i) => (
+          <img key={i} className="img" src={src} alt={`Team Profile ${i + 1}`} />
+        ))}
+      </ImageContainer>
+    </TeamHeroContainer>
+  );
+}
