+++ src/features/home/KnowledgeSection.jsx
@@ -1,13 +1,11 @@
-import React from 'react';
+import React, { useLayoutEffect, useRef } from 'react';
 import styled from '@emotion/styled';
-// import useBlurScrollEffect from '../../hooks/useBlurScrollEffect';
-// import KnowledgeOverlay from './KnowledgeOverlay';
+import gsap from 'gsap';
 import KnowledgeSectionLottie from './Knowledge/KnowledgeSectionLottie';
 import KnowledgeSectionList from './Knowledge/KnowledgeSectionList';
 
 const SectionContainer = styled.section`
   padding: 6rem 1.5rem;
-  background: linear-gradient(180deg, #f6f6f6 0%, #e6eaf5 100%);
 `;
 
 const SectionTitle = styled.h2`
@@ -17,7 +15,7 @@ const SectionTitle = styled.h2`
   margin-bottom: 1rem;
 
   @media (max-width: 768px) {
-    font-size: 1.2rem;
+    font-size: 1.8rem;
   }
 `;
 
@@ -32,22 +30,83 @@ const SectionSubtitle = styled.p`
   }
 `;
 
-const KnowledgeSection = () => {
-  // const sectionRef = useRef(null);
-  // const { blurAmount, opacity } = useBlurScrollEffect(sectionRef);
+const LottieWrapper = styled.div`
+  max-width: 1000px;
+  margin: 0 auto 3rem;
+  display: flex;
+  justify-content: center;
+`;
 
-  // blur가 일정 수준 이상일 때 오버레이 표시
-  // const showOverlay = blurAmount > 3;
+const ListWrapper = styled.ul`
+  max-width: 800px;
+  margin: 0 auto;
+  list-style: none;
+  padding: 0;
+`;
+
+export default function KnowledgeSection() {
+  const sectionRef = useRef(null);
+  const titleRef = useRef(null);
+  const lottieRef = useRef(null);
+  const listRef = useRef(null);
+
+  useLayoutEffect(() => {
+    const ctx = gsap.context(() => {
+      gsap.from(titleRef.current, {
+        opacity: 0,
+        y: 30,
+        duration: 0.8,
+        ease: 'power3.out',
+        scrollTrigger: {
+          trigger: titleRef.current,
+          start: 'top 80%',
+        },
+      });
+
+      gsap.from(lottieRef.current, {
+        scale: 0.8,
+        opacity: 0,
+        duration: 1,
+        ease: 'elastic.out(1, 0.5)',
+        scrollTrigger: {
+          trigger: lottieRef.current,
+          start: 'top 80%',
+        },
+      });
+
+      const items = listRef.current.querySelectorAll('.item');
+      items.forEach((item, i) => {
+        gsap.from(item, {
+          x: i % 2 === 0 ? -200 : 200,
+          opacity: 0,
+          ease: 'none',
+          scrollTrigger: {
+            trigger: item,
+            start: 'top 80%',
+          },
+        });
+      });
+    }, sectionRef);
+
+    return () => ctx.revert();
+  }, []);
+
   return (
-    <SectionContainer>
-      <SectionTitle>하루 한 조각, 지식을 채우는 시간</SectionTitle>
+    <SectionContainer ref={sectionRef}>
+      <SectionTitle ref={titleRef}>하루 한 조각, 지식을 채우는 시간</SectionTitle>
       <SectionSubtitle>
-        매일 이메일로 전달되는 짧지만 <br/>깊이 있는 지식 콘텐츠를 만나보세요.
+        매일 이메일로 전달되는 짧지만
+        <br />
+        깊이 있는 지식 콘텐츠를 만나보세요.
       </SectionSubtitle>
-      <KnowledgeSectionLottie />
-      <KnowledgeSectionList />
+
+      <LottieWrapper ref={lottieRef}>
+        <KnowledgeSectionLottie />
+      </LottieWrapper>
+
+      <ListWrapper ref={listRef}>
+        <KnowledgeSectionList />
+      </ListWrapper>
     </SectionContainer>
   );
-};
-
-export default KnowledgeSection;
+}
