+++ src/features/home/assets/hero6.avif
