+++ src/features/home/index.js
@@ -1,3 +1,4 @@
 export { default as HeroSection } from './HeroSection';
+export { default as AboutSection } from './AboutSection';
 export { default as KnowledgeSection } from './KnowledgeSection';
 export { default as PopularTopicSection } from './PopularTopicSection';
