--- src/features/team/TeamInfoSection.jsx
@@ -1,16 +1,0 @@
-import styled from '@emotion/styled';
-import TeamIntro from './TeamIntro';
-import TeamMember from './TeamMember';
-
-const TeamInfoContainer = styled.div``;
-
-const TeamInfoSection = () => {
-  return (
-    <TeamInfoContainer>
-      <TeamMember />
-      <TeamIntro />
-    </TeamInfoContainer>
-  );
-};
-
-export default TeamInfoSection;
