+++ src/features/team/hero/index.js
@@ -1,0 +1,1 @@
+export { default as TeamHeroSection } from './TeamHeroSection';
