+++ src/features/team/index.js
@@ -1,3 +1,3 @@
-export { default as TeamInfoSection } from './TeamInfoSection';
 export { default as TeamIntro } from './TeamIntro';
 export { default as TeamMember } from './TeamMember';
+export * from './hero';
