+++ src/pages/Home.jsx
@@ -1,10 +1,11 @@
-import { HeroSection, KnowledgeSection, PopularTopicSection } from '@features/home';
 import Layout from '@layout/Layout';
+import { HeroSection, AboutSection, KnowledgeSection, PopularTopicSection } from '@features/home';
 
 const Home = () => {
   return (
     <Layout>
       <HeroSection />
+      <AboutSection />
       <KnowledgeSection />
       <PopularTopicSection />
     </Layout>
