+++ src/pages/TeamInfo.jsx
@@ -1,11 +1,13 @@
 import React from 'react';
-import { TeamInfoSection } from '@features/team';
+import { TeamMember, TeamIntro } from '@features/team';
 import Layout from '@layout/Layout';
 
 const TeamInfo = () => {
   return (
     <Layout>
-      <TeamInfoSection />
+      {/* <TeamHeroSection /> */}
+      <TeamMember />
+      <TeamIntro />
     </Layout>
   );
 };
